completed all the tasks according to requirments

+++ src/components/CartCard.jsx
@@ -1,60 +1,43 @@
 import Swal from "sweetalert2";
 
 
-const CartCard = ({ cartProducts }) => {
+const CartCard = ({ users, setUsers }) => {
 
-    const { _id, image, name, brand, price } = cartProducts;
+    const { _id, image, name, brand, price } = users;
 
 
     const handleDelete = _id => {
         console.log(_id);
 
-        fetch(`http://localhost:5000/cart/${_id}`, {
-            method: 'DELETE'
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                fetch(`http://localhost:5000/cart/${_id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        if (data.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'The product has been deleted from the cart',
+                                'success'
+                            )
+                            const remaining = users.filter(user => user._id !== _id);
+                            setUsers(remaining);
+                        }
+                    })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                if (data.deletedCount > 0) {
-                    Swal.fire(
-                        'Deleted!',
-                        'The product has been deleted from the cart',
-                        'success'
-                    )
-                    // const remaining = users.filter(user => user._id !== _id);
-                    // setUsers(remaining);
-                }
-            })
-
-        // Swal.fire({
-        //     title: 'Are you sure?',
-        //     text: "You won't be able to revert this!",
-        //     icon: 'warning',
-        //     showCancelButton: true,
-        //     confirmButtonColor: '#3085d6',
-        //     cancelButtonColor: '#d33',
-        //     confirmButtonText: 'Yes, delete it!'
-        // }).then((result) => {
-        //     if (result.isConfirmed) {
-
-        //         fetch(`http://localhost:5000/cart/${_id}`, {
-        //             method: 'DELETE'
-        //         })
-        //             .then(res => res.json())
-        //             .then(data => {
-        //                 console.log(data);
-        //                 if (data.deletedCount > 0) {
-        //                     Swal.fire(
-        //                         'Deleted!',
-        //                         'The product has been deleted from the cart',
-        //                         'success'
-        //                     )
-        //                     // const remaining = users.filter(user => user._id !== _id);
-        //                     // setUsers(remaining);
-        //                 }
-        //             })
-        //     }
-        // })
     }
 
     return (
@@ -66,8 +49,8 @@ const CartCard = ({ cartProducts }) => {
                     <p className="font-extrabold text-xl text-emerald-600">Brand: {brand}</p>
                     <p className="font-extrabold text-xl text-lime-600">Price: {price}</p>
                     <div className="card-actions justify-end">
-                        <button onClick={() => handleDelete(_id)} className="btn btn-primary">Delete from Cart</button>
-                        <button className="btn btn-primary">Confirm Purchase</button>
+                        <button onClick={() => handleDelete(_id)} className="btn bg-emerald-500">Delete from Cart</button>
+                        <button className="btn bg-emerald-500">Confirm Purchase</button>
                     </div>
                 </div>
             </div>
